Simplify Imguibindings by using only the needed imgui functions
